refactor: modulariza em aplicações distintas
