+++ backend/evento/models.py
@@ -7,20 +7,6 @@ class MC:
         ('negado', 'Negado'),
         ('cancelado', 'Cancelado')
     )
-    metodo_pagamento = (
-        ('pix','Pix'),
-        ('boleto','Boleto'),
-        ('cartao','Cartão')
-    )
-    pagamento_status = (
-        ('pendente', 'Pendente'),
-        ('pago','Pago')
-    )
-    inscricao_status = (
-        ('pendente', 'Pendente'),
-        ('confirmada', 'Confirmada'),
-        ('cancelado', 'Cancelado')
-    )
     sexo = (
         ('M','Masculino'),
         ('F','Feminino'),
@@ -32,25 +18,6 @@ class MC:
         ('g', 'G'),
         ('gg', 'GG')
     )
-
-class localidade(models.Model):
-    cidade = models.CharField(max_length=100, blank=False, null=False)
-    uf = models.CharField(max_length=2, blank=False, null=False)
-
-class participante(models.Model):
-    nome = models.CharField(max_length=100, blank=False, null=False)
-    cpf = models.CharField(max_length=11, unique=True, blank=False, null=False)
-    email = models.EmailField(blank=False, null=False)
-    data_nascimento = models.DateField(blank=False, null=False)
-    telefone = models.CharField(max_length=15, blank=False, null=False)
-    rua = models.CharField(max_length=100, blank=False, null=True)
-    numero = models.CharField(max_length=10, blank=False, null=True)
-    bairro = models.CharField(max_length=100, blank=False, null=True)
-    localidade = models.ForeignKey(localidade, on_delete=models.CASCADE, related_name='participantes', default=1)
-    
-class organizador(models.Model):
-    valor = models.DecimalField(max_digits=10, decimal_places=2, blank=False, null=False, default=0)
-    participante = models.OneToOneField(participante, on_delete=models.CASCADE, related_name='organizadores')
 
 class evento(models.Model):
     nome = models.CharField(max_length=100, blank=True, null=True)
@@ -64,8 +31,8 @@ class evento(models.Model):
     quantInscAtual = models.PositiveIntegerField(blank=False, null=False, default=0)
     valorInsc = models.DecimalField(max_digits=6, decimal_places=2, blank=False, null=False)
     status = models.CharField(choices=MC().evento_status, max_length=10, default='pendente')
-    organizador = models.ForeignKey(organizador, on_delete=models.CASCADE, related_name='eventos')
-    localidade = models.ForeignKey(localidade, on_delete=models.CASCADE, related_name='eventos', default=1)
+    organizador = models.ForeignKey('usuarios.organizador', on_delete=models.CASCADE, related_name='eventos')
+    localidade = models.ForeignKey('localidades.localidade', on_delete=models.CASCADE, related_name='eventos', default=1)
     imagem = models.ImageField(upload_to='', null=True, blank=True)
 
 class categoria(models.Model):
@@ -84,19 +51,3 @@ class item(models.Model):
     nome = models.CharField(max_length=30, null=True, blank=True)
     tamanho = models.CharField(choices=MC().tamanhos, max_length=2, null=True, blank=True)
     kit = models.ManyToManyField(kit, related_name='itens')
-
-class inscricao(models.Model):
-    dataInsc = models.DateField(auto_now=True)
-    status = models.CharField(choices=MC().inscricao_status, default='pendente')
-    evento = models.ForeignKey(evento, on_delete=models.CASCADE, related_name='inscricoes')
-    participante = models.ForeignKey(participante, on_delete=models.CASCADE, related_name='inscricoes', null=True)
-    kit = models.ForeignKey(kit, on_delete=models.CASCADE, related_name='inscricoes', null=True)
-    categoria = models.ForeignKey(categoria, on_delete=models.CASCADE, related_name='inscricoes', null=True)
-    
-class pagamento(models.Model):
-    status = models.CharField(choices=MC().pagamento_status, max_length=10, default='pendente')
-    valor = models.DecimalField(max_digits=10, decimal_places=2, blank=False, null=False, default=0)
-    metodoPagamento = models.CharField(choices=MC().metodo_pagamento, max_length=20, null=False, blank=False)
-    dataPagamento = models.DateField(auto_now=True)
-    inscricao = models.OneToOneField(inscricao, on_delete=models.CASCADE, related_name='pagamentos')
-
